+++ src/nanonext.h
@@ -189,7 +189,7 @@ typedef struct nano_saio_s {
   nng_ctx *ctx;
   nng_aio *aio;
   void *cb;
-  int msgid;
+  int id;
 } nano_saio;
 
 typedef struct nano_cv_s {
@@ -255,7 +255,6 @@ extern SEXP nano_HeadersSymbol;
 extern SEXP nano_IdSymbol;
 extern SEXP nano_ListenerSymbol;
 extern SEXP nano_MonitorSymbol;
-extern SEXP nano_MsgidSymbol;
 extern SEXP nano_ProtocolSymbol;
 extern SEXP nano_ResolveSymbol;
 extern SEXP nano_ResponseSymbol;